pyproject.toml
[tool.poetry]
name = "plantilla-proyecto-python"
version = "0.5.0"
description = "Plantilla para organizar proyectos incluyendo las mejores prácticas"
authors = ["Alejandro Perez Londoño <perezl.alejandro@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pydantic-settings = "^2.2.1"
loguru = "^0.7.0"
icecream = "^2.1.3"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
ruff = "^0.3.7"
pip-audit = "^2.7.2"
pre-commit = "^3.7.0"

[tool.poetry.group.tests.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocstrings = "^0.24.3"
mkdocstrings-python = "^1.9.2"
mkdocs-autorefs = "^1.0.1"
mkdocs-material = "^9.5.17"
mkdocs-include-markdown-plugin = "^6.0.5"

[tool.poetry.group.kernels]
optional = true

[tool.poetry.group.kernels.dependencies]
spyder-kernels = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = 3.11
files = ["src/**/*.py", "tests/**/*.py"]
plugins = [ "pydantic.mypy" ]
exclude = [".venv"]
no_namespace_packages = true
allow_redefinition = false
check_untyped_defs = true
disallow_any_explicit = false
disallow_any_generics = true
disallow_untyped_calls = true
ignore_errors = false
ignore_missing_imports = true
ignore_missing_imports_per_module = true
implicit_reexport = false
local_partial_types = true
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_no_return = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true

[tool.ruff]
line-length = 88
src = ["src", "test"]
target-version = "py311"

[tool.ruff.lint]
select = [
  "A",   # builtins
  "B",   # bugbear
  "COM", # commas
  "C90", # mccabe
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "I",   # isort
  "N",   # naming
  "PL",  # pylint
  "UP",  # pyupgrade
  "S",   # bandit
]
ignore = [
  "COM812", # trailing comma missing ...
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # asserts allowed in tests...
]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["D104"]

[tool.ruff.lint.isort]
known-first-party = ["configuracion", "logger"]

[tool.ruff.lint.mccabe]
max-complexity = 6

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
